--- AssignmentAlert.js
@@ -1,4 +1,0 @@
-import React from 'react';
-import {Alert, Text, TouchableOpacity, StyleSheet} from 'react-native';
-
-// const assignmentAlert = (message) => ( )
