+++ android/app/src/main/java/com/newworld/MainApplication.java
@@ -24,9 +24,7 @@ public class MainApplication extends Application implements ReactApplication {
     @Override
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-          new CookieManagerPackage()
-          new MapsPackage()
+          new MainReactPackage()
       );
     }
