+++ index.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
-import Root from './src/Root';
+import Root from './src/components/robots/RobotInstructions';
 
 AppRegistry.registerComponent('NewWorld', () => Root);
