+++ package.json
@@ -9,11 +9,11 @@
   "dependencies": {
     "eslint": "^3.19.0",
     "react": "16.3.1",
-    "react-native-cookies": "^3.3.0",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.0",
     "react-native": "0.54",
-    "react-native-maps": "^0.21.0"
+    "react-native-cookies": "^3.3.0",
+    "react-native-maps": "^0.21.0",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0"
   },
   "devDependencies": {
     "babel-jest": "22.4.4",
