+++ src/Root.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Provider } from "react-redux";
 
 import configureStore from "./store";
-import App from "./App";
+import App from "./components/robots/RobotInstructions";
 
 const Root = () => (
   <Provider store={configureStore()}>
